+++ common/libs/libcamera/sc8830/isp_calibration/src/isp_cali_interface.c
@@ -144,7 +144,7 @@ int32_t ISP_Cali_GetLensTabs(struct isp_addr_t img_addr,
 		fread(lnc_tmp_buf, 1, length, fp);
                 fclose(fp);
 		tg_lnc.y_addr = (uint32_t)lnc_map.map[x][y].param_addr;
-		mg_lnc.y_addr = lnc_tmp_buf;
+		mg_lnc.y_addr = (uint32_t)lnc_tmp_buf;
 		mr_lnc.y_addr = (uint32_t)lnc_tab_buf;
 		final_lnc.y_addr = (uint32_t)lens_tab;
